requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
environs==11.0.0
fastapi==0.112.0
fastapi_cors==0.0.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
marshmallow==3.21.3
openai==1.40.2
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
